Add list of all locales
This is a pre-amble to managing translations with a human-friendly list of locales to choose from. The reason why the list of locales is in a .json file rather than a part of the database is simply because it's not a list that should be modified very often. The list is copied from Wikipedia with the addition of the "en" entry (which is the default locale for the Translations module).