prevent creating duplicate or empty translations
set a defaultLanguage config variable which is the default for the app
as a whole. This is used in a few places to look for all the
translations definitions.

use beforeValidate to prevent creating empty or identical translations
to the translation inherited.

Use the defaultLangage whenever paginating data to have a full list of
translation keys. Otherwise, for example, a newly created language will
be empty, or a partially translated langauge not show all keys

When creating new translation entries _always_ create them in the
default locale