need to extend the uniqe index.
Plural rules will mean we end up storing:

    id	key			plural_id	plural_case
	1   1 Item		null		null
	2	%d Items	1			0
	3	%d Items	1			1
	4	%d Items	1			3
	5	%d Items	1			4
	6	%d Items	1			5
	7	%d Items	1			6

So, ensure we can store that

revised schema.

for singular translations - allow to mark if they have a plural.

This will make it easier to administer plurals in the backend. II
envision a manage plural form button, which would open a form to allow
defining the various forms - that'd only be relevant if the plural rules
are complex enough for "it's the plural form" to be confusing such as:

	Arabic
	Czech
	Icelandic
	Polish
	Russian

The above list is not exhaustive, these are the languages which IFF we
use them in the future we need to have the means to handle their plural
rules and a means to provide a meaningful interface.