remove tests for nolonger supported replacements
these are ambiguous now that the singular translation also receives the
number - if we shifted the translation on as the first argument as
before the single translations don't work - they get the number. Without
array-shifting the count into the argument array, the plural
translations don't work. Rather than some monkey patch to get it to work
- just remove the tests for unwanted usage.